fixed: Make use of known max texture size for GL textures, and make it unsigned.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport@22576 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/RenderSystemGL.cpp
@@ -396,7 +396,7 @@ void CRenderSystemGL::CalculateMaxTexturesize()
     }
   }
 
-  CLog::Log(LOGINFO, "GL: Maximum texture width: %d", (int) m_maxTextureSize);
+  CLog::Log(LOGINFO, "GL: Maximum texture width: %u", m_maxTextureSize);
 }
 
 void CRenderSystemGL::GetViewPort(CRect& viewPort)
